Update FolderServiceImpl.java
修复文件夹内容为空时程序异常的bug

+++ FileManagement/src/main/java/com/jothub/service/impl/FolderServiceImpl.java
@@ -72,10 +72,12 @@ public class FolderServiceImpl implements FolderService {
         List<Integer> fileIdsToDelete = fileMapper.getFileIdsByFolderIds(folderIdsToDelete);
 
         // 删除文件记录和文件物理存储
-        for (String filePath : fileMapper.getFilePathsByIds(fileIdsToDelete)) {
-            File file = new File(filePath);
-            if (file.exists()) {
-                file.delete();
+        if(!fileIdsToDelete.isEmpty()) {
+            for (String filePath : fileMapper.getFilePathsByIds(fileIdsToDelete)) {
+                File file = new File(filePath);
+                if (file.exists()) {
+                    file.delete();
+                }
             }
         }
